error handler refactoring
Error handler methods are adjusted and the routes folder is deleted.
