+++ app/main.py
@@ -1,11 +1,12 @@
 import fastapi
 
 from app.routes import metrics
-from app.config import get_settings
+from app.utils.config import get_settings
 from logging import getLogger
+from app.utils import context_vars
+from app.utils.exception_handlers import http_exception_handler, validation_exception_handler
+from app.utils.middleware import log_requests
 from starlette.exceptions import HTTPException as StarletteHTTPException
-from uuid import uuid4
-from app.services.utils import context_vars
 
 settings = get_settings()
 settings.configure_logging()
@@ -25,47 +26,10 @@ app = fastapi.FastAPI(
     docs_url=None if settings.env.lower() == "prod" else "/docs",
 )
 
+app.middleware("http")(log_requests)
 
-@app.middleware("http")
-async def log_requests(request: fastapi.Request, call_next):
-    request_id = str(uuid4())
-    request_id_context.set(request_id)
-
-    logger.info(
-        f"Method: {request.method} - URL: {request.url}",
-        extra={"request_id": request_id},
-    )
-
-    response = await call_next(request)
-    logger.info(
-        f"Response status: {response.status_code }",
-        extra={"request_id": request_id},
-    )
-    return response
-
-
-@app.exception_handler(StarletteHTTPException)
-async def http_exception_handler(request, exc):
-    logger.error(
-        f"HTTP Exception: {exc} - Path: {request.url} - Method: {request.method}",
-        extra={"request_id": request_id_context.get()},
-    )
-    return fastapi.responses.JSONResponse(
-        exc.detail, status_code=exc.status_code
-    )
-
-
-@app.exception_handler(fastapi.exceptions.RequestValidationError)
-async def validation_exception_handler(request, exc):
-    logger.error(
-        f"Validation error: {exc.errors()} - Path: {request.url} - Method: {request.method}",
-        extra={"request_id": request_id_context.get()},
-    )
-    return fastapi.responses.JSONResponse(
-        status_code=422,
-        content={"detail": str(exc.errors())},
-    )
-
+app.add_exception_handler(StarletteHTTPException, http_exception_handler)
+app.add_exception_handler(fastapi.exceptions.RequestValidationError, validation_exception_handler)
 
 app.include_router(metrics.router)
 # TODO: disable swagger on production: https://fastapi.tiangolo.com/tutorial/metadata/#docs-urls
