+++ requirements.txt
@@ -7,3 +7,4 @@ rio-tiler>=6.6.1,<6.7.0
 rasterio>=1.3.9,<1.3.10
 geopandas>=1.0.0,<1.0.1
 shapely>=2.0.3,<2.0.4
+starlette>=0.37.2,<0.38.0
